# src/main/resources/api_doc.yml
swagger: '2.0'
info:
  description: Api Documentation
  version: '1.0'
  title: Api Documentation
  termsOfService: urn:tos
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0
host: localhost:8080
basePath: "/"
...
paths:
...
  "/api/category/add":
    post:
      tags:
      - category-controller
      summary: addCategories
      operationId: addCategoriesUsingPOST
      consumes:
      - multipart/form-data
      produces:
      - "*/*"
      parameters:
      - name: id
        in: query
        description: id
        required: true
        type: integer
        format: int64
      - name: title
        in: query
        description: title
        required: true
        type: string
      - name: description
        in: query
        description: description
        required: true
        type: string
      - name: photo
        in: formData
        description: photo
        required: true
        type: file
      responses:
        '200':
          description: OK
          schema:
            type: object
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  "/api/category/categories":
    get:
      tags:
      - category-controller
      summary: getCategories
      operationId: getCategoriesUsingGET
      produces:
      - "*/*"
      responses:
        '200':
          description: OK
          schema:
            type: object
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  "/api/category/delete":
    post:
      tags:
      - category-controller
      summary: delete
      operationId: deleteUsingPOST
      consumes:
      - application/json
      produces:
      - "*/*"
      parameters:
      - name: id
        in: query
        description: id
        required: true
        type: integer
        format: int64
      responses:
        '200':
          description: OK
          schema:
            type: object
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
...